requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
PyYAML>=6.0.2
pandas>=2.2.2
tenacity>=8.5.0
tqdm>=4.66.4
regex>=2024.5.15
langchain>=0.2.12
langgraph>=0.2.14
langchain-google-genai>=2.0.2
pydantic>=2.8.2
streamlit>=1.36.0

# New dependencies for multi-source approach
playwright>=1.40.0
trafilatura
spacy>=3.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
sentence-transformers>=2.2.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=4.9.0

# UI enhancements
plotly>=5.18.0
openpyxl>=3.1.0

